# openapi.yaml
openapi: 3.0.3
info:
  title: BarberPOS Backend API
  version: 0.1.0
servers:
  - url: https://api.barberpos.cloud
paths:
...
  /auth/register:
    post:
      summary: Register with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/google:
    post:
      summary: Login/Register with Google (idToken)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/staff:
    post:
      summary: Login staff/employee (by phone or email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/forgot-password:
    post:
      summary: Request reset code (dummy 1234)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Code sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          code: { type: string }
  /auth/reset-password:
    post:
      summary: Reset password (dummy success if token==1234)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
  /auth/change-password:
    post:
      summary: Change password for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
...
  /attendance/checkin:
    post:
      summary: Check-in attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
                  format: int64
                employeeName:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
  /attendance/checkout:
    post:
      summary: Check-out attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
                  format: int64
                employeeName:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
  /attendance:
    get:
      summary: Get attendance for month
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: employeeName
          required: true
          schema:
            type: string
        - in: query
          name: month
          required: true
          schema:
            type: string
            example: 2025-01
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            employeeId: { type: integer, format: int64, nullable: true }
                            employeeName: { type: string }
                            date: { type: string }
                            checkIn: { type: string, nullable: true }
                            checkOut: { type: string, nullable: true }
                            status: { type: string }
                            source: { type: string }
...
  /logs:
    get:
      summary: List activity logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ActivityLog'
    post:
      summary: Create activity log (sync)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityLogCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
                          id: { type: integer, format: int64 }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: integer }
        status: { type: string }
    ApiResponse:
      type: object
      required: [status, data]
      properties:
        status: { type: string, enum: [success, error] }
        message: { type: string }
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
    RegisterRequest:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        phone: { type: string }
        address: { type: string }
        region: { type: string }
        role: { type: string, example: manager }
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    StaffLoginRequest:
      type: object
      properties:
        phone: { type: string }
        email: { type: string }
        name: { type: string }
        pin: { type: string, description: "Employee PIN" }
    GoogleLoginRequest:
      type: object
      properties:
        idToken: { type: string }
        email: { type: string }
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        region: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: string }
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        phone: { type: string }
        address: { type: string }
        region: { type: string }
        isGoogle: { type: boolean }
...
    ActivityLogCreateRequest:
      type: object
      properties:
        title: { type: string }
        message: { type: string }
        actor: { type: string }
        type: { type: string, enum: [info, warning, error] }
        timestamp: { type: string, format: date-time }
    ActivityLog:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        message: { type: string }
        actor: { type: string }
        type: { type: string }
        timestamp: { type: string, format: date-time }
        synced: { type: boolean }
...